requirements.txt
asn1crypto==0.24.0
bcrypt==3.1.4
cffi==1.11.5
cryptography==2.2.2
Django==1.10.6
django-haystack==2.6.1
enum34==1.1.6
Fabric3==1.14.post1
idna==2.6
ipaddress==1.0.19
jieba==0.39
Markdown==2.6.11
paramiko==2.4.1
Pillow==5.1.0
pyasn1==0.4.2
pycparser==2.18
Pygments==2.2.0
PyNaCl==1.2.1
pytz==2018.3
six==1.11.0
Whoosh==2.7.4
